Change DebugPolicy to non-static DebugPolicySource
